Updated windows events with renamed package
