+++ cmd/kawa/config_unix.go
@@ -4,10 +4,10 @@
 package main
 
 import (
-	"github.com/runreveal/flow"
-	"github.com/runreveal/flow/internal/sources/journald"
-	"github.com/runreveal/flow/internal/sources/syslog"
-	"github.com/runreveal/flow/internal/types"
+	"github.com/runreveal/kawa"
+	"github.com/runreveal/kawa/internal/sources/journald"
+	"github.com/runreveal/kawa/internal/sources/syslog"
+	"github.com/runreveal/kawa/internal/types"
 	"github.com/runreveal/lib/loader"
 	"golang.org/x/exp/slog"
 	// We could register and configure these in a separate package
@@ -17,29 +17,31 @@ import (
 )
 
 func init() {
-	loader.Register("syslog", func() loader.Builder[flow.Source[types.Event]] {
+	loader.Register("syslog", func() loader.Builder[kawa.Source[types.Event]] {
 		return &SyslogConfig{}
 	})
-	loader.Register("journald", func() loader.Builder[flow.Source[types.Event]] {
+	loader.Register("journald", func() loader.Builder[kawa.Source[types.Event]] {
 		return &JournaldConfig{}
 	})
 }
 
 type SyslogConfig struct {
-	Addr string `json:"addr"`
+	Addr        string `json:"addr"`
+	ContentType string `json:"contentType"`
 }
 
-func (c *SyslogConfig) Configure() (flow.Source[types.Event], error) {
+func (c *SyslogConfig) Configure() (kawa.Source[types.Event], error) {
 	slog.Info("configuring syslog")
 	return syslog.NewSyslogSource(syslog.SyslogCfg{
-		Addr: c.Addr,
+		Addr:        c.Addr,
+		ContentType: c.ContentType,
 	}), nil
 }
 
 type JournaldConfig struct {
 }
 
-func (c *JournaldConfig) Configure() (flow.Source[types.Event], error) {
+func (c *JournaldConfig) Configure() (kawa.Source[types.Event], error) {
 	slog.Info("configuring journald")
 	return journald.New(), nil
 }
